DON-524 – make components standalone
WIP: first steps towards defining all requirements
independently and removing modules
