+++ src/app/app.server.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { FlexLayoutServerModule } from '@angular/flex-layout/server';
-import { ServerModule, ServerTransferStateModule } from '@angular/platform-server';
+import { ServerModule } from '@angular/platform-server';
 
 import { AppModule } from './app.module';
 import { AppComponent } from './app.component';
@@ -10,7 +10,6 @@ import { AppComponent } from './app.component';
     AppModule,
     FlexLayoutServerModule, // Must come after AppModule, see https://github.com/angular/flex-layout/blob/master/guides/SSR.md
     ServerModule,
-    ServerTransferStateModule,
   ],
   bootstrap: [AppComponent],
 })
