+++ src/app/campaign-details/campaign-details.component.spec.ts
@@ -4,7 +4,6 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatProgressBarModule } from '@angular/material/progress-bar';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { MatTabsModule } from '@angular/material/tabs';
-import { BrowserTransferStateModule } from '@angular/platform-browser';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { RouterTestingModule } from '@angular/router/testing';
 
@@ -21,7 +20,6 @@ describe('CampaignDetailsComponent', () => {
     TestBed.configureTestingModule({
       declarations: [ CampaignDetailsComponent, CampaignDetailsCardComponent, TimeLeftPipe ],
       imports: [
-        BrowserTransferStateModule,
         HttpClientTestingModule,
         MatIconModule,
         MatProgressBarModule,
