+++ src/app/donation-start/donation-start-matching-expired-dialog.component.ts
@@ -1,8 +1,13 @@
 import { Component } from '@angular/core';
+import { MatDialogModule } from '@angular/material/dialog';
 
 @Component({
+  standalone: true,
   selector: 'app-donation-start-matching-expired-dialog',
   templateUrl: 'donation-start-matching-expired-dialog.html',
+  imports: [
+    MatDialogModule,
+  ],
 })
 export class DonationStartMatchingExpiredDialogComponent {
   constructor() {}
