+++ src/app/navigation/navigation.component.ts
@@ -3,6 +3,7 @@ import { Router, NavigationStart, RouterEvent } from '@angular/router';
 import { filter } from 'rxjs/operators';
 
 @Component({
+  standalone: true,
   selector: 'app-navigation',
   templateUrl: './navigation.component.html',
   styleUrls: ['./navigation.component.scss'],
