+++ src/app/promoted-campaigns/promoted-campaigns.component.ts
@@ -1,12 +1,17 @@
 import { Component, Input } from '@angular/core';
 
 import { Campaign } from '../campaign.model';
+import { CampaignPromoCardComponent } from '../campaign-promo-card/campaign-promo-card.component';
 import { environment } from '../../environments/environment';
 
 @Component({
+  standalone: true,
   selector: 'app-promoted-campaigns',
   templateUrl: './promoted-campaigns.component.html',
-  styleUrls: ['./promoted-campaigns.component.scss']
+  styleUrls: ['./promoted-campaigns.component.scss'],
+  imports: [
+    CampaignPromoCardComponent,
+  ]
 })
 export class PromotedCampaignsComponent {
   @Input() campaign1: Campaign;
