+++ src/main.server.ts
@@ -8,4 +8,4 @@ if (environment.productionLike) {
 
 export { AppServerModule } from './app/app.server.module';
 
-export { renderModule } from '@angular/platform-server';
+export { renderApplication } from '@angular/platform-server';
